requirements.txt
# <INSTALLED_ADDONS>  # Warning: text inside the INSTALLED_ADDONS tags is auto-generated. Manual changes will be overwritten.
django==2.2.11
gunicorn==19.7.1
psycopg2-binary==2.8.4
django-storages==1.8
django-amazon-ses==2.1.1
django-heroku==0.3.1
django-cacheops==4.2
django-extensions==2.2.5
django-nose==1.4.6
django-positions==0.6.0
django-proxy==1.2.1
django-redis==4.10.0
django-replicated==2.6.1
django-rq-scheduler==1.1.3
django-rq==2.1.0
nose-exclude==0.5.0
nose==1.3.7
python-dotenv==0.10.3
rq-scheduler==0.9.1
rq==1.1.0
s3transfer==0.2.1
sentry-sdk==0.14.3
watchdog==0.9.0
boto3==1.10.29
botocore==1.13.29
django-compressor==2.4
django-autoslug==1.9.7
django-summernote==0.8.11.6
django-uuslug==1.2.0

asn1crypto==0.24.0
beautifulsoup4==4.7.1
bleach==3.1.0
bottle==0.12.16
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
defusedxml==0.5.0
diff-match-patch==20181111
django-ckeditor==5.6.1
django-crequest==2018.5.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-elasticsearch-dsl==6.4.1
django-elasticsearch-dsl-drf==0.18
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-mdeditor==0.1.13
django-model-utils==3.1.2
django-nine==0.2.2
django-notifications-hq==1.5.0
django-ranged-response==0.2.0
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.9.4
drf-haystack==1.8.5
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.12.1
idna==2.8
jdcal==1.4
jieba==0.39
jsonfield==2.0.2
Markdown==3.1
odfpy==1.4.0
openpyxl==2.6.0
Pillow==5.4.1
pycparser==2.19
Pygments==2.3.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-memcached==1.59
pytz==2017.2
PyYAML==5.1.2
redis==3.3.7
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.2.4
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.25.3
webencodings==0.5.1
WeRoBot==1.9.0
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0